Added the ability to do changesets only
The default behavior is to accept all baselines, then accept all changesets.

BnS appears to have baselines that include prior baselines, so this would result in one huge commit for the single baseline that's available, then a handful of commits after.

This allows one to use changesets only, which should be manageable with BnS's shorter history (compared to the SFI project)